# openapi.yaml
openapi: 3.0.3
info:
  title: HLS Streaming Server API
  description: |
    Complete REST API for managing HLS streaming channels, media libraries, and dynamic playlists.

    ## Features
    - Multi-channel HLS streaming with FFmpeg
    - Dynamic playlist generation via schedule blocks
    - Jellyfin-style library management
    - Media bucket organization
    - EPG/XMLTV generation
    - IPTV M3U playlist export
    - Session-based authentication
    - Virtual time tracking for 24/7 simulation

    ## Authentication
    The API supports two authentication methods:
    1. **API Key** (Legacy): Pass via `X-API-Key` header or `apiKey` query parameter
    2. **Session Token** (Modern): Pass via `Authorization: Bearer <token>` header or `sessionToken` cookie

    Most endpoints require authentication except for streaming and public EPG endpoints.

  version: 1.0.0
  contact:
    name: HLS Streaming Server
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://your-server.com
    description: Production server
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
    CookieAuth:
      type: apiKey
      in: cookie
      name: sessionToken
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        message:
          type: string
      required:
        - success
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
      required:
        - success
        - error
...
paths:
...
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register first admin user
      description: Register the first admin user (only works if no users exist)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  minLength: 3
                password:
                  type: string
                  minLength: 8
                email:
                  type: string
                  format: email
              required:
                - username
                - password
      responses:
        '200':
          description: User registered successfully
          headers:
            Set-Cookie:
              schema:
                type: string
                example: sessionToken=abc123; Path=/; HttpOnly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Validation error or users already exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Login with credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/logout:
    post:
      tags: [Authentication]
      summary: Logout and invalidate session
      security:
        - BearerAuth: []
        - CookieAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/auth/me:
    get:
      tags: [Authentication]
      summary: Get current user info
      security:
        - BearerAuth: []
        - CookieAuth: []
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          userId:
                            type: string
                            format: uuid
                          username:
                            type: string

  # === CHANNELS ===
...